packages/bots/axiom-assist-bot/mcp-server/requirements.txt
mcp>=1.21.1
google-cloud-firestore